third_party/blink/web_tests/accessibility/clickable-with-handler-on-window.html

S<!DOCTYPE HTML>
<html>
<body>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>

<div id="x" role="button">abc</div>
<script>
window.addEventListener('click', () => { console.log('hi'); });

test(function(t) {
    assert_true(accessibilityController.accessibleElementById('x').isClickable);
}, "An ARIA widget descendant of a clickable window is clickable.");
</script>

</body>
</html>
